Fixed imports in cosmwasm/enclaves/shared/cosmos-types and tweaks to package dependencies
