+++ cosmwasm/enclaves/execute/Cargo.toml
@@ -19,7 +19,7 @@ debug-print = []
 test = []
 
 # This annotation is here to trick the IDE into showing us type information about this crate.
-# we alwats compile to the "sgx" target, so this will always be false.
+# We always compile to the "sgx" target, so this will always be false.
 # when compiling to the "sgx" target, we pull this from the target root with an "extern crate" directive
 [target.'cfg(not(target_env = "sgx"))'.dependencies]
 sgx_types = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
