+++ cosmwasm/enclaves/ffi-types/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2018"
 default = []
 build_headers = ["cbindgen", "thiserror"]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
 derive_more = "0.99"
