+++ cosmwasm/enclaves/shared/cosmos-types/Cargo.toml
@@ -4,23 +4,25 @@ version = "1.2.4"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-#[features]
-#SGX = []
-#default = ["prost"]
+# This annotation is here to trick the IDE into showing us type information about this crate.
+# We always compile to the "sgx" target, so this will always be false.
+# when compiling to the "sgx" target, we pull this from the target root with an "extern crate" directive
+[target.'cfg(not(target_env = "sgx"))'.dependencies]
+sgx_tstd = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = [
+  "backtrace"
+] }
 
 [dependencies]
-serde = { git = "https://github.com/mesalock-linux/serde-sgx", features = ["derive"] }
-
-protobuf = "2.25.2"
-log = "0.4.14"
-derive_more = "0.99"
-sha2 = "0.8.1"
-
-prost = {git = "https://github.com/mesalock-linux/prost-sgx", rev = "cd3103a6d45cf7a43b6c1c5e4223428097d1c547",default-features = false,  features = ["prost-derive"]}
-
+cosmos_proto = { path = "../cosmos-proto" }
 enclave-ffi-types = { path = "../../ffi-types" }
 enclave_crypto = { path = "../crypto" }
-cosmos_proto = { path = "../cosmos-proto" }
 enclave_cosmwasm_types = { path = "../cosmwasm-types" }
+
+log = "0.4.14"
+serde = { git = "https://github.com/mesalock-linux/serde-sgx", features = ["derive"] }
+derive_more = "0.99"
+
+protobuf = "2.25.2"
+sha2 = "0.8.1"
+
+prost = { git = "https://github.com/mesalock-linux/prost-sgx", rev = "cd3103a6d45cf7a43b6c1c5e4223428097d1c547", default-features = false,  features = ["prost-derive"] }
