--- cosmwasm/enclaves/shared/cosmos-types/src/errors.rs
@@ -1,7 +1,0 @@
-use derive_more::Display;
-
-#[derive(Debug, Display)]
-pub enum SignatureError {
-    VerificationError,
-    ParsingError,
-}
