+++ cosmwasm/enclaves/shared/cosmos-types/src/traits.rs
@@ -1,11 +1,9 @@
-use crate::types::CosmosAddress;
+use enclave_cosmwasm_types::types::CanonicalAddr;
 
 // https://github.com/tendermint/tendermint/blob/v0.33.3/crypto/crypto.go#L22
 pub trait CosmosAminoPubkey: PartialEq {
     /// derive the canonical address for this public key
-    fn get_address(&self) -> CosmosAddress;
+    fn get_address(&self) -> CanonicalAddr;
     /// Serialize this public key to the legacy Amino format
     fn amino_bytes(&self) -> Vec<u8>;
-    // Verify that `sig` was generated by this public key, for the content in `bytes`
-    // fn verify_bytes(&self, bytes: &[u8], sig: &[u8]) -> Result<(), CryptoError>;
 }
