+++ cosmwasm/enclaves/shared/cosmwasm-types/Cargo.toml
@@ -4,18 +4,20 @@ version = "1.2.4"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
+# This annotation is here to trick the IDE into showing us type information about this crate.
+# We always compile to the "sgx" target, so this will always be false.
+# when compiling to the "sgx" target, we pull this from the target root with an "extern crate" directive
+[target.'cfg(not(target_env = "sgx"))'.dependencies]
 sgx_tstd = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = [
   "backtrace"
 ] }
+
+[dependencies]
+enclave-ffi-types = { path = "../../ffi-types" }
+
+log = "0.4.8"
 serde = { git = "https://github.com/mesalock-linux/serde-sgx", features = ["derive"] }
 serde_json = { git = "https://github.com/mesalock-linux/serde-json-sgx" }
-
-enclave-ffi-types = { path = "../../ffi-types" }
-log = "0.4.8"
-bech32 = "0.7.2"
-sgx_types = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 derive_more = "0.99"
+bech32 = "0.7.2"
 base64 = { rev = "dc7389e10817b078f289386b3b6a852ab6c4c021", git = "https://github.com/mesalock-linux/rust-base64-sgx" }
