+++ cosmwasm/enclaves/shared/crypto/Cargo.toml
@@ -4,30 +4,30 @@ version = "1.2.4"
 authors = ["Cashmaney <itzik@enigma.co>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 # This annotation is here to trick the IDE into showing us type information about this crate.
-# we alwats compile to the "sgx" target, so this will always be false.
+# We always compile to the "sgx" target, so this will always be false.
 # when compiling to the "sgx" target, we pull this from the target root with an "extern crate" directive
 [target.'cfg(not(target_env = "sgx"))'.dependencies]
-sgx_types = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 sgx_tstd = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = [
   "backtrace"
 ] }
 
 [dependencies]
-log = "0.4.8"
-aes-siv = "=0.6.2"
-enclave-ffi-types = { path = "../../ffi-types" }
-ring = { git = "https://github.com/mesalock-linux/ring-sgx", tag = "v0.16.5" }
+sgx_types = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 sgx_trts = { rev = "a37ffb9449ba6d5b6e4a9d586bbab864ae732269", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
 
-ripemd160 = "0.9.1"
-secp256k1 = "0.19.0"
-x25519-dalek = { version = "=1.2.0", default-features = false, features = [
-  "u64_backend"
-] }
-sha2 = "0.8.1"
+enclave-ffi-types = { path = "../../ffi-types" }
+
+log = "0.4.8"
 derive_more = "0.99"
 lazy_static = "1.4"
 hex = "0.4.2"
+
+ring = { git = "https://github.com/mesalock-linux/ring-sgx", tag = "v0.16.5" }
+sha2 = "0.8.1"
+ripemd160 = "0.9.1"
+secp256k1 = "0.19.0"
+aes-siv = "=0.6.2"
+x25519-dalek = { version = "=1.2.0", default-features = false, features = [
+  "u64_backend"
+] }
