+++ cosmwasm/enclaves/test/Cargo.toml
@@ -4,7 +4,5 @@ version = "0.1.0"
 authors = ["Reuven Podmazo <reuven@enigma.co>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 cosmwasm-sgx-vm = { path = "../../packages/sgx-vm", features = ["iterator", "enclave-tests"] }
