+++ cosmwasm/enclaves/xargo/sgx_tstd/cargo.toml
